backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
PyPDF2
asgiref
sqlparse
tzdata
PyJWT
